# book-Network-UI/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BookResponse {
  archived?: boolean;
  authorName?: string;
  cover?: Array<string>;
  id?: number;
  isbn?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopis?: string;
  title?: string;
}