# book.Network/src/main/java/com/rabebBou/book/feedback/FeedBackResponse.java
package com.rabebBou.book.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedBackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;



}